fix score_graph so it doesn't crash
