+++ src/R_causality/R_causality.h
@@ -24,8 +24,8 @@ int          edge_to_int(const char *edge);
 const char * edge_to_char(int edge);
 int          node_to_int(const char *node, const char **nodes);
 
-int * calculate_edges_ptr(SEXP Graph);
 void calculate_edges_from_cgraph(struct cgraph *cg, SEXP graph);
+struct dataframe prepare_df(SEXP Df, SEXP States);
 struct cgraph * cgraph_from_causality_graph(SEXP Graph);
 SEXP causality_graph_from_cgraph(struct cgraph *cg, SEXP Nodes);
 SEXP create_causality_graph(int n_nedges, int n_nodes, SEXP Nodes);
